Reverted back from FP16 to FP32

+++ lib/test/evaluation/tracker.py
@@ -342,9 +342,9 @@ class Tracker:
 
         with torch.no_grad():
             device = 'cpu'
-            template = template.to(device).type(torch.float16)
-            search = template.to(device).type(torch.float16)
-            seq_input = template.to(device).type(torch.float16)
+            template = template.to(device).type(torch.FloatTensor)
+            search = template.to(device).type(torch.FloatTensor)
+            seq_input = template.to(device).type(torch.FloatTensor)
             dummy_input = (template, search, seq_input)
             print(f"{template=}, {template.shape}")
             print(f"{search=}, {search.shape}")
